pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.63.1",
    "python-dotenv>=1.0.1",
    "requests>=2.28.0",
    "pillow>=9.0.0",
    "moviepy==1.0.3",
    "numpy==1.24.3",
    "decorator==4.4.2",
    "imageio==2.31.1",
    "imageio-ffmpeg==0.4.8",
    "tqdm==4.65.0",
    "proglog==0.1.9",
    "telegram>=0.0.1",
    "python-telegram-bot>=20.0",
]
